Started to add landscape
Added a landscape VBO, hex-style grid all in one triangle strip!
Made a shader that reads height data from the red channel of an image for the
height of landscape.
Made use of the Camera class
Tidied up the javascript a little
Made UI.static an object that gets automatically saved when quitting, and
reloaded when starting again, so you can save stuff there.
Made UI.static.autorun, which is an array, each member gets call()ed if it's a function,
or eval()ed otherwise at startup.